Spring Boot 2.x基础教程：使用JdbcTemplate访问MySQL数据库

+++ 2.1.x/chapter3-1/src/test/java/com/didispace/chapter31/Chapter31ApplicationTests.java
@@ -7,10 +7,12 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
+import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.RequestBuilder;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.transaction.annotation.Transactional;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
@@ -41,14 +43,14 @@ public class Chapter31ApplicationTests {
         userSerivce.create("Linda", 17);
 
         // 查数据库，应该有5个用户
-        Assert.assertEquals(5, userSerivce.getAllUsers().intValue());
+        Assert.assertEquals(5, userSerivce.getAllUsers());
 
         // 删除两个用户
         userSerivce.deleteByName("Tom");
         userSerivce.deleteByName("Mike");
 
         // 查数据库，应该有5个用户
-        Assert.assertEquals(3, userSerivce.getAllUsers().intValue());
+        Assert.assertEquals(3, userSerivce.getAllUsers());
 
     }
